requirements.txt
fastapi
uvicorn
pydantic
langchain-google-genai
langchain-community
python-dotenv